Move rewritten function to autograd_function.py
