+++ torch/autograd/function.py
@@ -561,7 +561,7 @@ class Function(_SingleLevelFunction):
 
             return bound_args.args
 
-        is_setup_ctx_defined = is_setup_context_defined(cls.setup_context)
+        is_setup_ctx_defined = _is_setup_context_defined(cls.setup_context)
         if is_setup_ctx_defined:
             args = bind_default_args(cls.forward, *args, **kwargs)
 
@@ -585,17 +585,8 @@ class Function(_SingleLevelFunction):
         return (ctx._autograd_function_id,)
 
 
-def is_setup_context_defined(fn):
+def _is_setup_context_defined(fn):
     return fn != _SingleLevelFunction.setup_context
-
-
-def autograd_function_forward_rewritten(original_forward, original_setup_context):
-    def new_forward(ctx, *args, **kwargs):
-        output = original_forward(*args, **kwargs)
-        original_setup_context(ctx, args, output)
-        return output
-
-    return new_forward
 
 
 def once_differentiable(fn):
